+++ mPULSE Project/mPULSE/RenderingEngine/ShopScreenInfo.h
@@ -27,7 +27,7 @@ public:
     int selectedMenuIndex;
     int selectedItemIndex;
 
-    float obsCost;
+    int obsCost;
 
     int newShipTexture;     //Need to tell which texture to use for new models;
